requirements.txt
pydantic
PyYAML
pytest